move `prbuilder` under `cmd` for better project structure consistency
